requirements.txt
geopandas>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
folium>=0.14.0
osmnx>=1.6.0
networkx>=3.1
scikit-learn>=1.3.0
statsmodels>=0.14.0
libpysal>=4.9.0
esda>=2.5.0
mgwr>=2.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
census>=0.8.19
us>=3.0.0
dash>=2.14.0
jupyter>=1.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
